+++ aerial_robot_control/CMakeLists.txt
@@ -66,9 +66,9 @@ add_library(flight_control_pluginlib
   src/control/under_actuated_controller.cpp
   src/control/under_actuated_lqi_controller.cpp
   src/control/under_actuated_tilted_lqi_controller.cpp
-  src/control/nmpc_controller.cpp
-  src/control/base/mpc_solver.cpp)
-target_link_libraries(flight_control_pluginlib ${catkin_LIBRARIES} control_utils acados ${PROJECT_SOURCE_DIR}/include/aerial_robot_control/control/base/nmpc_body_rate/c_generated_code/libacados_ocp_solver_qd_body_rate_model.so)
+  src/nmpc/nmpc_controller.cpp
+  src/nmpc/unactuated_body_rate/mpc_solver.cpp)
+target_link_libraries(flight_control_pluginlib ${catkin_LIBRARIES} control_utils acados ${PROJECT_SOURCE_DIR}/include/aerial_robot_control/nmpc/unactuated_body_rate/c_generated_code/libacados_ocp_solver_qd_body_rate_model.so)
 add_dependencies(flight_control_pluginlib  ${PROJECT_NAME}_gencfg)
 
 ### flight navigation
