+++ aerial_robot_control/include/aerial_robot_control/nmpc/unactuated_body_rate/c_generated_code/libacados_ocp_solver_qd_body_rate_model.so
