+++ aerial_robot_control/include/aerial_robot_control/nmpc/unactuated_body_rate/c_generated_code/qd_body_rate_model_cost/qd_body_rate_model_cost_y_e_fun_jac_ut_xt.c
