+++ aerial_robot_control/include/aerial_robot_control/nmpc/unactuated_body_rate/qd_body_rate_model_acados_ocp.json
