+++ aerial_robot_control/src/nmpc/README.md
@@ -1,0 +1,7 @@
+# Naming rule
+
+Folder: object name + `_` + model output
+
+e.g. underactuated_body_rate means the robot is underactuated (quadrotor) and the model output is body rate.
+
+# Change log
