+++ aerial_robot_control/src/nmpc/unactuated_body_rate/mpc_solver.cpp
@@ -3,7 +3,7 @@
 // Created by lijinjie on 23/10/22.
 //
 
-#include "aerial_robot_control/control/base/mpc_solver.h"
+#include "aerial_robot_control/nmpc/unactuated_body_rate/mpc_solver.h"
 
 MPC::MPCSolver::MPCSolver()
 {
